AAC: Refactored code, moved mutagen specific code to mutagenext
